Fixed|libappfw|GLTextComposer: Showing a portion of tabbed content
The visible line range was not taken properly into account when
laying out tabbed content for drawing. Also, when lines outside the
range are released, their metadata is still kept intact even though
the rasterized version is released from the atlas. This allows the
text layout to still be calculated.