Refactor|SectorCluster|Client: Continued cleaning up Shard preparation
Now that a Shard's DrawList spec can be configured post construction
and calculation of the per-primitive texture transform matrices are
deferred until draw time; it is no longer worthwhile to branch the
preparation logic according to whether the geometry is destined for
the skymasked list or not.

Todo: The next step is revise how dynamic light/shadows are projected
to map surfaces. Given that the dynamics change far more frequently
than the map geometry, we need to be able to do this independently.
One question is whether it makes sense to continue drawing the lights
with multitexturing (would mean some non-trivial bookkeeping in order
to reformat Shards dynamically).