Renderer|DrawList|Shard: Replaced DrawList::Spec::GeomGroup with Shard::Type
Rather than determine at Shard construction time which DrawList the
geometry will be written to, it makes more sense to classify Shards
according to their logical type/usage. The GeomGroup is no longer a
part of the DrawList::Spec, instead these are now used to identify
the DrawList(s) and their respective texture hashes (if used). The
GeomGroup for a given Shard is now automatically determined when a
compatible list is requested from the DrawLists collection.

Another benefit of doing it this way is that one is completely free
to reconfigure Shards dynamically (so long as they haven't already
been added to a DrawList, that is).