Fixed|Hexen|GameSession: Always update the internal.save when revisiting a map
Somewhere down the line the logic for updating the old "reborn" slot
was confused resulting in explicit logic to not update internal.save
when revisiting a previously visited map. This is of course wrong as
if the player dies and they then attempt to load their progress they
will instead be taken to the map which they last left (in the current
hub).

I also inlined the P_RemoveAllPlayerMobjs() logic into GameSession
as its unlikely we'll need to do this anywhere else.