Refactor|libdoomsday|FS1|Wad: Reimplemented Wad's internal lump LUT with LumpIndex
Public access to the LumpIndex is provided as Wad is derived from it.
This pushes the high level design of the Wad component is much closer
to that of the ZipArchive / Archive model used by FS2.

The idea here is adapt the old FS1 components to use similar design
to FS2 and thereby easing the transition to the new file system.