Refactor|libdoomsday|FS1|Zip: Reimplemented Zip with LumpIndex, Entry and LumpFile
This component now follows the high level design of FS1 Wad.

Next step: Invert the model of both FS1 Wad and Zip such that each
derives from PathTreeT with the LumpIndex as a subcomponent. It is
important to note that in the case of Wad, that the design cannot
follow the ZipArchive / Archive model exactly because Archive does
not currently allow for multiple files with the same name.