WadMapConverter|Id1Map: Heuristic recognition of id Tech 1 format maps
Rather than rely on the game to provide the name of the map "header"
lump, traverse the WAD lump index looking for lump name sequences of
expected map data. When a sequence of named lumps is found the id of
the map can then be determined and the data lumps conformance checked
according to the expected size for the determined format.

This means that Id1Map can now be used to locate maps without relying
on hints from the game (meaning, it is now possible to locate them
before they are requested by the currently loaded game).