All Games: Command "warp" now also accepts a map URI
One can now specify the map to warp to using a map URI rather than a
logical map number (and episode, depending on the game mode).
For example; warp MAP01; warp E1M1

Although this means that it is now possible to warp to maps that are
not part of the defined map progression, this functionality should be
considered experimental and is intended for development purposes only.

Hexen note:
Map URIs are not translated according to any "warptrans" mappings in
MAPINFO. Rather, this is the actual URI of the map (e.g., warp MAP40
will take you to Dark Crucible, whereas warp 40 won't work).