Refactor|libcore|FS: Added a Node base class for files
filesys::Node is the base class for File. It stores the most basic
members: name and parent. It also has overridable logic for following
a path from a node.

This is only a minimal file system node class; at some point in the
future, the management of child nodes could be moved here from Folder.