Scripting|FS|libcore: Added built-in functions 'File' and 'globals'
The File() function returns a File's info record. The info records of
all File instances now have the superclass Core.File. This superclass
provides methods for manipulating files: currently it has methods for
locating files via a relative path and reading the contents of a file
into a string or a Block.

globals() returns the process's global namespace as a record.