libcommon: G_ComposeMapUri() and G_MapNumberFor() are now deprecated
Map references composed of a logical episode and map number pair are
a historical legacy that should only be used when necessary for
compatibility reasons.

As of this commit episode numbering is completely separate from map
identifiers and one should no longer assume that this information can
be extracted from them. The assumption being that the episode number
is defined in MapInfo.