libcore|FS: Assets in packages
A package asset is a Record in the package metadata plus optionally some
corresponding files in the package.

The "/packs" folder now links to packages also by asset identifiers, not
just package identifiers. This allows easily locating the package
currently providing a particular asset.

App has convenient methods for locating package assets.

filesys::AssetObserver is a utility for getting notified when assets
matching a regexp are available or unavailable.