Fixed|libcommon: Do not attempt player reborns if no game session is in progress
If loading a saved game fails at a specific point in the process, it
is presently possible that the game will enter an undefined state.

Dealing with such a failure is straight forward when no game is loaded.
Doing so when a game is loaded is a bit more involved, however.