Refactor|API|Thinker: Base thinker relocated to libdoomsday
Removed the 'inStasis' member of the base thinker struct with a set of
flags. Thinker_SetStasis() was moved out of the Thinker API to
libdoomsday. The Thinker API version was bumped.

Also added a C++ wrapper for managing a POD thinker. This is the
first step toward a C++ thinker class hierarchy. The wrapper handles
ownership of the new private thinker data instance, and will be used
for duplicating and copying thinkers. The wrapper supports both
standard and memory zone allocations.