Cleanup|libcommon|Menu: Removed unused 'parameters' arg from menu::Widget action callbacks
This is effectively a "widget event" mechanism. There is no real use
for routing arguments in this way now that the menu widgets have been
updated to use an object-oriented model. Indeed most of these callbacks
are used to implement widget-internal behaviors, rather than to action
game menu logics specifically.