Refactor|Client: Added MobjThinker and applied ThinkerT
MobjThinker is a specialized ThinkerT that uses the game's mobj size
for the thinker. This is now used by the client's delta processor and
sound emitters for line, plane and sector.

MobjThinker is also used for allocating new mobjs. The private data
of a mobj is created JIT before the mobj thinks for the first time.
The private data is freed after the thinker is removed from the map,
or when the map is destroyed.

Debug: Since there is no actual private data yet, a simple DebugCounter
object is used to keep tabs on all the private objects, to make sure
they are all created and deleted appropriately. This will be replaced
with a private data object constructed by a factor.