Refactor|Client|World: Converted ClPlaneMover to be ThinkerData-based
Converted ClPlaneMover to use the new ThinkerData private data to
try out how this works out in practice. ClPlaneMover is now
object-oriented, and handles its construction on its own.

Since the private data of a thinker is deleted as part of the thinker,
there is no need for special ClPlaneMover related code in Map.

All ClPlaneMover related functionality was removed from Map. There is
no need for map-wide indexing of movers; better to let Plane know
if there is currently a ClPlaneMover acting on it. The end result is
more elegant and efficient.