IdTech1Converter|libcommon: Moved MAPINFO reading logic to IdTech1Converter
From here we'll translate the MAPINFO definitions into DED. Doing it
this way we can isolate the messy logic needed for compatibility and
present this data to the games via the DED database. (It also means
that this data will be available from Doomsday Script).

Sadly it won't be possible to do the translation in a single pass,
however.