DED: Added 'Episode' definitions
Episode definitions are used to define metadata (such as the title of
the episode), how the episode is presented in the game menu and most
importantly, the starting map on which to begin.

The previously hardcoded default episode definitions have were moved
to DED files in the game plugins' resource pack. This means they can
now be customized by mods.

Todo: Translate Hexen "warp numbers" when used as map references.