Fixed|libdoomsday|DEDRegister: Handling initially empty lookup key values
If a lookup key is not given a value when initially added, it will
still be observed so that when the value is given, it will be added
to the appropriate lookup.

However, this assumes that only the latest record is being modified
(otherwise the lookup may not point to the correct entry).