qmake|Windows: Write built binaries directly to "distrib/products"
Now DESTDIR and DLLDESTDIR are used to write all built binaries directly
to the distrib folder. This cuts down on the time needed for running
the install target. Installing remains necessary, though, as it copies
all the supporting files.

windeployqt is now run by postbuild, making it unnecessary to manually copy
any Qt libraries to the distrib folder.