Console|libcommon: Enhanced "leavemap" command
leavemap now accepts an optional 'exit' argument, to specify the name
of the exit to lookup from the map progression defined for the current
episode and map. This gives mod authors a convenient way to test out
the map progression without having to find the associated trigger.

It is also now possible to leavemap in Hexen.