libcommon|GameSession: Switched GameSession to use textual episode identifiers
A valid episode id is required in order to begin a new game session.
The GPT_GAME_STATE packet now includes the current episode id. Always
force a new game session when the episode specified to the "warp" set
of commands differs from the current. Added convenience methods to
common::GameSession for returning the current episode id and map URI.

Added G_EpisodeTitle as a convenient way of retrieving an episode
title by episode id and improved the map title banner to include it.