libhexen: Animated sky setup is now handled by the engine
Previously it was necessary to do the sky setup on game side because
the config was partly defined in the Hexen specific MAPINFO data.

As this data is to be translated into DED during game init it is now
unnecessary to do the sky setup on gameside.

Plus some cleanup.

Todo: Re-implement sky layer scrolling on engine side.