IdTech1Converter|MapInfoTranslator: ZDoom-compatible interpretation of cluster numbers
In ZDoom-MAPINFO the episode concept is interpreted as a cluster whose
ClusterDef is not attributed with the "is-hub" property.

As of this commit all supported games and all the mods I have tested
appear to be working as expected other than a remaining issue with the
incompatible music song/track interpretation. (ZDoom does not abstract
music interpretation to group lump number, CD track, etc...).