Fixed|Resources: Model skin selection error
When (re)building the #skin => Texture map one must ensure that the
indices uses in the MD2/DMD internal skin table are respected, else
the selector mechanism will choose the wrong skin. This means that
"duplicated" slots must map to the same Texture and missing textures
result in a NULL mapping for that skin number.

Issue #1871