Fixed|Console|All Games: Don't begin a game state change message if waiting for a response
This is possible if the user tries to circumvent the message response
prompt by opening the console and entering a command that triggers a
new message.

Todo: These game state messages should be considered modal - opening
of the console should be prevented while a message which requires a
response is active.