Fixed|XG: Surface tint color changed when triggering XG 'plane_material' lines
Due to a refactoring oversight back in 2007, the logic of the XG
plane_material line class was changed to always change the surface
tint color when triggered (default is black).

Evidently, however, back when this was implemented originally the
necessary "Change Color" flag was omitted, so this was added also.

IssueID #965