Fixed|Client|Windows: Potential crash if closing window during startup
When game resources are being located, games become current for short
periods of time (kludge in Games::locateStartupResources()). If a
"quit" command happened to be executed at this time, a crash could
occur in the console command execution, presumably if the game suddenly
was removed in another thread.

As a workaround, closing the window is disallowed during busy mode.

IssueID #1895