Console|libdoomsday: Rudimentary access to de::Config variables
This commit introduces a simple mechanism that allows the user to
interactively access de::Config variables as if they were traditional
cvars. The allowed operations are checking the value of the variable
(having it printed in the console), or setting the value (number or
text only).

This is not a long term solution, though. When the console is based on
Doomsday Script, this mechanism should become unnecessary and obsolete.
(Although some form of mapping between traditional cvar identifiers and
Config variables is still needed.)