DED|libdoomsday: Added if (not) custom condition to Mods directive
Presently supported with Map Info definitions only. This conditional
allows one to instruct the DED parser to apply the modification only
when the "custom" status of the definition matches.

For example, to reassign the music of E1M1 to a different song only
if the associated Map Info definition has not yet been modified by
an add-on:

Map Info mods "E1M1" if not custom { Music = "e1m1"; }