libgui: GLFramebuffer uses multisampled textures
Now multisampling is done by rendering directly into a multisampled
texture in GLFramebuffer. There is no need for a proxy target.
A shader is used to manually combine the samples when the multisampled
color texture is drawn to the non-multisampled back buffer.

Also added a few more queried limits to GLInfo.