Refactor|Map|World: Implement important map element/object iterations with C++11 lambdas
Things are now looking significantly cleaner, and should also perform
a little better in an optimized Release build.

There are further iteration mechanisms which would benefit from using
a lambda based approach. However, we've now reached the end of the
"unstable" development phase for 1.15, so these will have to wait.