Fixed|Input|Client|libgui: Handling mouse wheel events (Qt 5)
In Qt 5, the pixel and angle deltas are provided separately. The former
is for trackpads while the latter is for actual wheels.

Also fixed how the wheel steps are passed to the game's input handling,
so that only wheel step events are handled and not the pixel based
scrolling.

IssueID #1909