UI|Home|Client: Delete savegames via the Game Selection menu/dialog
The saved sessions are now represented in a visually different way in
Game Selection; the visual distinction makes it easier to see the
difference between saved sessions and the other kinds of sessions.

The saved sessions have an X icon that allows deleting the savegame.

Made text styling more consistent in the Home screen info popups
('dimmed' color for paths, bold 'accent' for headings).

There are two variants of the close icon, one without a ring and one
with a ring. These are supposed to be used depending on the context
(e.g., with the ring in menus).

IssueID #1771