Fixed|Doom: Updated IDMUS cheat handling re custom episodes
The IDMUS cheat now respects custom episodes while also attempting
to interpret the given arguments in a vanilla compatible manner.
Obviously this is somewhat limited by the inflexible mechanics of
the cheat mechanism and the fixed/predefined semantics.

These changes mean that it is no longer possible to play music that
is not associated with a map using the IDMUS cheat. (However, these
can still be played using the "playmusic" command, which is a far
more flexible method in any case).

IssueID #1914