libappfw: Improvements for widget background blurring
Instead of each and every transparent widget doing a full redraw of
the entire view for the blur, now the Style can provide a common
source for the blurred background. This can improve rendering
performance significantly.

Added a GuiWidget attribute that allows a widget to draw its blurred
background into a texture even though the widget itself is not
visible.