UI|Client|Performance: Reduced redundant background blurring
The client window already has a task bar blurring widget that gets
used during the tutorial. The same blurred content is now used in
Ring Zero for all widgets that have a blurred background.

In the future, the same can be done when a game is loaded, when game
rendering is more optimized for less redraw overhead.