libcommon|Menu: Skip Episode selection menu when navigating backwards
If only one playable episode is available we shouldn't bother the
user to manually select it. In such a case, automatically select the
episode and proceed to next/previous page.

Note that the demo/shareware versions are exempted because they use
this menu to prompt the user to buy the full version.

IssueID #1911