Fixed|Map Renderer|Client|Debug: Crash attempting to sort vissprites
If rend-dev-freeze 1 was set prior to starting a map, no vissprites
were generated during the first frame, meaning visSpriteP == NULL,
thus when subtracted from visSprites yielded an completely incorrect
count of the number of vissprites in use.

Consequently, R_SortVisSprites() trashed whatever happened to be in
memory following the fixed-size visSprites array and utterly confused
itself when trying to sort the results. Ultimately causing a crash
when it attempted to deference an invalid pointer.