Fixed|Renderer: Z-buffering of sky models
3D models won't look right without Z buffering, however they must not
interfere with world geometry. This commit enables Z buffering for
sky models and clears the depth buffer afterwards.

Excessive clears of the depth buffer could be avoided by using a
separate depth buffer for the sky, however the sky stencil should be
drawn there as well. This could be a topic for a future refactoring.

IssueID #1922