UI|Client|Default Style|libappfw: Item alignment in popup menus
This commits cleans up the layout of popup menu items to reduce visual
noise. The idea is that items are aligned by their text, rather than
the widget's left edge. Possible item images hang on the left.

Also added a "Close Log" action to the Console menu that dismisses
the log. "Clear Log" was moved so it isn't mistaken for the way to
close the log.