libcommon: Removed current episode ID from the map banner (log)
On second thoughts, lets not include this info in the map banner so
as not to potentially confuse the user when manually warping.

Once a game session has begun, the episode is "locked in" at this
point. Any future map changes are interpreted as a continuation of
the current episode.

Consequently printing the current episode ID here is potentially
confusing, when switching to using map URIs.

For example, "warp 3 8" is interpreted as "Begin a new game on episode
3 and then warp to map-with-warp-number 8". Whereas "warp E3M8" is
interpreted as (simply) "warp to map-with-warp-number 8".

(Note that if both an episode ID and a map URI is specified, a new
game session will begin, as previously).