Fixed|Multiplayer|Client: Cleaning up client state when leaving a game
The client was not properly cleaning its state when leaving a game.
Now received packets are discarded and the postfx shader is reset to
"none".

If one was kicked from an MP game due to version conflict, the old
buffered packets caused a segfault when the client continued handling
them after connecting to a new server.

IssueID #1954
IssueID #1971