UI|Client|Task Bar: Choosing the IWAD folder
Added a menu item in the app menu to select the primary IWAD folder
that overrides all others. This is saved persistently in Config.

Whenever the path selection dialog is accepted, game resources are
located again (even if the path doesn't change so there's a way for
the user to recheck if files have been moved in the native file
system).