Fixed|UI|Client: Using the correct frame for busy transitions
Previously there was an arbitrary 0.5 second threshold for determining
if the captured busy frame was still fine for a transition (checked when
a new request for a busy frame occurred).

This arbitrary threshold was removed and the busy frame is only
discarded when the regular game is running without a busy transition.

Also did some cleanup: removing obsolete/commented code.

IssueID #1939