Fixed|libappfw: DocumentWidget content overflowing the right edge
Fixes a layout issue with DocumentWidget where it was possible that
tabbed lines would extend beyond the predetermined bounds for the
text content. This was occurring for instance in the About -> GL popup
when the first set of tabbed lines were shorter than the second set.

Todo for later: Improve FontLineWrapping to precalculate tab stop
alignment so this can be known before vertex generation.