libgui|Canvas|GLFramebuffer: Changing sample count at runtime
When the Canvas frame buffer's sample count changes, the Canvas's
native window will be recreated with the updated format.

Care must be taken when reshowing the native window, because an
expose event might occur synchronously, leading to recreation of widget
background blur FBs, which in turn would invalidate (destroy) some of
the audience of the default sample count property.