Refactor|Renderer: Restructured texture => surface projection
Revised the high-level model for the projection of textures onto
world Surfaces, making use of C++11 lambdas to improve SoC and
better locality. RenderSystem now has ownership of the projected
texture lists.

Todo for later: There is no need to expose the projected texture
list instances to users of RenderSystem.