Fixed|Hexen: Death Wyvern gets stuck
The main trigger for the Wyvern getting stuck was that it chased the
dead Ettin that sometimes falls on the lava. If the Ettin is too
close to the central pillar, the Wyvern's requirement to get close
enough to the target was impossible.

Now the Wyvern just forgets about the Ettin after it has died.

Fixed the incorrect use of FIX2FLT (we're using coord_t),
addressed compiler warnings about calling abs() on an unsigned int,
and cleaned up the code a little.

IssueID #1634