CMake|Assimp|Unix: Link to and deploy libassimp
The generator expressions for accessing the right library files to link
to and install seem to be different on each platform. However, note that
it is a bit unorthodox to access the build products of an external
project via the CMake target object. The current setup exists because
the "libassimp" interface target can either represent an Assimp
installation somewhere in the file system, or the build products of the
Assimp that gets built as part of Doomsday.