Fixed|CMake: CMake configs for SDK libraries
When building the complete project, the config files (*Config.cmake)
of the SDK libraries were being accessed from the install directory,
even though that should only occur when using a pre-installed SDK.

This caused many of the "net.dengine" packages to go missing from the
installation, because in a clean build the library configs were not
found and thus the applications didn't know which packages to deploy.

Now the library configs are in their own directory under
doomsday/cmake/, where they can be found during the build. Only
when DENG_SDK_DIR is specified by the user, the configs are searched
from an external location.

On the Mac, the packages are now installed only as part of the SDK
component, since usually they get bundled with the application.