CMake|CPack: Removed Assimp packages, only install enabled components
Now that Assimp is part of the build, CMake was automagically detecting
that its components should be part of the Doomsday packaging. These were
removed because they were breaking the packaging.

Now Doomsday will only package the enabled components. The server and
savegametool are always installed as part of the "client" component rather
than the optional "tools" component.