Refactor|Renderer|Client: Moved map geometry buffer writes out of DrawList
Tasking DrawList with performing map geometry buffer writes results
in an awkward bottleneck that not only encumbers the DrawList class
with this responsibility but also forces the users of it to first
build the geometry locally beforehand.

Todo (Optimize): Now that geometry writes are performed independently
of DrawList the actual geometry generation can be re-written to avoid
the need for temporary storage and passing the actual geometry data
down the line (instead only buffer used, the type(s) of primitive and
their vertex indices need passing).

IssueID #2077