API: Removed 'sprNames' table from InternalData
None of the current game plugins currently access this data so lets
not export it. Bumped DE_API_INTERNAL_DATA to ver 2.

Todo: Doomsday should automatically generate Sprite definitions from
the sprite frame resources in WAD files (similarly to Materials), if
placed and named according to id Tech 1 conventions. Game plugins and
Doomsday Scripts will be able to access the definitions via the Defs
database. Authors will be able to define new Sprites and/or modify
the IWAD originals using .ded files.

If necessary, ResourceSystem could build a runtime representation of
this data for use with the renderer (probably won't be; most of the
relevant selection logic can be encapsulated in a new defn::Sprite
accessor and asset/resource caching is handled at another level).