Model Renderer|GL2: Moved rendering to ModelRenderer; model mirroring
Assimp uses a different handedness in the coordinate system, which
means the model needs mirroring when it is drawn.

The model drawing routine was moved to its rightful place in the
ModelRenderer class, which was already managing part of the state for
drawing models. This now also includes the per-model cull mode.

Added a "mirror" parameter for the model info that flips the side
vector (perpendicular to both the defined up and front vectors).