CMake|MSVC: Build products are runnable within build directory (on Windows)
The output paths for binaries were changed so that the files form a
products tree inside the build directory. Also, when CMake is run, it
copies the required 3rd party libraries and data packages to the build
products directories.

This allows running Doomsday.exe directly from within the build directory
after finishing a build.

The changes did not affect the INSTALL step, which is still done when
creating distribution packages.