Log|libappfw|Client: Added PrivilegedLogWidget (developer utility)
Often when working on a specific piece of code, it is easiest to just
use qDebug to print debug messages because they can be easily seen
among the other messages.

Doomsday's Log should provide a better way to handle this use case, so
now there are Privileged entries. Instead of qDebug, you can now use
LOG_WIP() in a similar fashion, and the entries will show up in the
special overlaid log in the top left side of the screen.

LOG_WIP() entries will appear normally in doomsday.out and the standard
output, however they are omitted from the runtime console log widget.