Fixed|Model Renderer: Eye position for specular lighting
Previously only the eye direction was passed to the shader, but
for accurate results, the eye direction has to be calculated separately
for each vertex. Now only the eye position in local space is given
to the shader.

Also fixed specular lighting from light sources on the wrong side of
surfaces.