Model Renderer: Animation priorities and loop re-triggering
It is now possible to set a priority value for each animation sequence
(default: 1). Lower priority animations cannot start when a higher
priority animation is running. Instead, they get queued for starting
when the higher priority animation has finished.

When a looping animation has finished its loop, it will retrigger
animations according to its state, which may cause a different variant
of the animation to be started.