Scripting|Info|libcore: No exception when parsing an embedded script
ScriptLex can now be told to stop parsing statements when it encounters
a mismatched closing }. This occurs normally when a Doomsday Script
block is parsed inside an Info document, because the Info parser has
no way of knowing how long the embedded script is unless it actually
parses it using ScriptLex.

Previously ScriptLex just blindly tried to read statements until it
hit an exception.