libappfw|PopupWidget: Smarter popups: don't eat clicks, auto-flip direction
This commit makes dealing with popups more fluent and reduces the
number of clicks needed to work with the UI, because clicking outside
a popup will automatically close the popup and allow the clicked
widget to react normally.

This removes a kludge that previously resent mouse clicks to specific
widgets when a popup was closing. This was not ideal because events
should normally be propagated to all widgets, not just the clicked
ones.

Popups also by default now reverse their opening direction if there
is more space in the other direction.