OpenGL|libgui: Check for EXT_framebuffer_object; more robust FBO setup
Should check for EXT_framebuffer_object rather than the ARB extension,
because the ARB extension was added for OpenGL 3 and we are still
working with OpenGL 2.1.

If the render-to-texture FBO fails, and the 24+8 depth-stencil texture
fails, try separate render buffers for depth and stencil (using the
default bit depths). If that also fails, attempt to create a simple
FBO with render buffers only. This will break parts of the renderer,
but allow the engine to start.