Fixed|Model Renderer: Inadvertently borrowing other objects' bone transformations
ModelDrawable objects are shared between all the mobjs of the same type,
so if one of them doesn't have anything to update the bone
transformations with, it will reuse the transformations of the
previously drawn object.

Now the mobj animator keeps the last animation sequence, even if it's
finished, so there's always something to animate the bones with.