Fixed|Model Renderer|World: Defer triggering of state animations
A mobj may change states multiple times per tick. For instance, an
Ettin's attack could be restarted immediately during the same state
when the previous attack has ended and the Ettin returns to chase.

The animations are now triggered based on what is the mobj state in
effect in the end of a tick.