Refactor|Audio: Moved server side audio::System functionality to ServerApp
ServerApp itself now provides a tracking mechanism for logical sounds
that are playing in the Server side's view of the world.

audio::System now exists solely on Client side.

Todo: Continue whittling away at the server side audio functionality.
It may well be that this mechanism actually belongs in the Server's
network domain (it clearly has very little to do with audio per-say).