Server: Replaced sample cache with a WaveformData database
The server has no need for caching sound samples (as they will never
be played locally) and because purging was never preformed the cache
was actually being used as a glorified database in any case.

Todo: The server should be clearing this data when the game changes!
Todo: Consolidate logic for locating, processing and resampling of
sound sample resources (for both Client and Server) and relocate to
the ResourceSystem.