Refactor|Audio|Client: Fully encapsulate LogicSound management in audio::System
Users of the audio::System should not need to be concerned with the
mechanics of managing logical sound tracking. Rather than ask users
of this class to interact with a separate logical sound manager, it
is cleaner to enumerate the available SoundStages and have the user
specify the stage on which to perform sound start/stop actions and
is-playing queries. The audio::System can infer from the specified
sound stage how/when to adjust the logical sound management.