Model Renderer: Rendering pass specific shaders/variables
A model definition can now specify which shader to use (via
"render.shader" in the asset metadata).

Additionally, each render pass definition can now include variables
and the name of the shader to be used during the pass.

It is possible that shaders use different texture mapping layouts
in their vertex data. ModelRenderer enforces compatible layouts
(i.e., the order of the texture coordinates for the various texture
maps must be the same for the ones that are in use). This ensures
that the model's vertex buffers can remain static during rendering.